OP-319 - The way the controller is overwritten has been changed
